pyproject.toml
[project]
name = "information-technology-project-management"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "langchain-openai>=0.2.11",
    "matplotlib>=3.9.3",
    "neo4j>=5.27.0",
    "networkx>=3.4.2",
    "nltk>=3.9.1",
    "numpy<2",
    "ollama>=0.4.2",
    "openai>=1.56.1",
    "pip>=24.3.1",
    "polars>=1.16.0",
    "pypdf>=5.1.0",
    "python-docx>=1.1.2",
    "python-dotenv>=1.0.1",
    "pyvis>=0.3.2",
    "ragas>=0.2.6",
    "ragchecker>=0.1.9",
    "sentence-transformers>=3.3.1",
    "streamlit-option-menu>=0.4.0",
    "streamlit>=1.40.2",
    "transformers>=4.46.3",
    "google-generativeai>=0.8.3",
    "groq>=0.13.0",
    "aioboto3>=13.2.0",
    "lightrag-hku>=1.0.6",
    "nano-vectordb>=0.0.4.3",
]
